Move runnables to separate file
