+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "frogware_fcxqm"
-version = "0.0.1"
+version = "0.0.2"
 authors = [
   { name="Mike Wahl", email="mike.wahl@colorado.edu" },
 ]
