+++ src/frogware_fcxqm/Gui_Controller.py
@@ -1,21 +1,24 @@
 import sys
 import PyQt5.QtWidgets as qt
 import PyQt5.QtCore as qtc
-from .window import Ui_MainWindow
-from . import plottablefunctions as plotf
-import numpy as np
-from .error import Ui_Form
 import PyQt5.QtGui as qtg
 import gc
 import threading
 import scipy.integrate as scint
 import matplotlib.pyplot as plt
-from pylablib.devices.Thorlabs.kinesis import list_kinesis_devices
+import numpy as np
+
+from .runnables import UpdateMotorPositionRunnable, UpdateSpectrumRunnable, Signal
+from .window import Ui_MainWindow
+from . import plottablefunctions as plotf
+from .error import Ui_Form
 from .hardware_comms.kinesis import ThorlabsKinesisMotor
 from .hardware_comms.ocean import OceanOpticsSpectrometer
 from .hardware_comms.device_interfaces import LinearMotor, Spectrometer, SpectrometerAverageException, StageOutOfBoundsException, SpectrometerIntegrationException
 from .hardware_comms.utilities import T_fs_to_dist_um, dist_um_to_T_fs
+
 from seabreeze.spectrometers import Spectrometer as ooSpec
+from pylablib.devices.Thorlabs.kinesis import list_kinesis_devices
 
 # will be used later on for any continuous update of the display that lasts more
 # than a few seconds
@@ -24,15 +27,6 @@ pool = qtc.QThreadPool.globalInstance()
 # global variables
 tol_um = 0.1  # 100 nm
 edge_limit_buffer_mm = 0.0  # 1 um
-
-# Signal class to be used for Runnable
-
-
-class Signal(qtc.QObject):
-    started = qtc.pyqtSignal(object)
-    progress = qtc.pyqtSignal(object)
-    finished = qtc.pyqtSignal(object)
-
 
 # Popup error window
 class ErrorWindow(qt.QWidget, Ui_Form):
@@ -1155,90 +1149,6 @@ class CollectSpectrogram:
         # increment spectrogram collection index, and iterate again
         self.n += 1
         self.step_one()
-
-
-class UpdateMotorPositionRunnable(qtc.QRunnable):
-    def __init__(self, motor: LinearMotor, event_to_clear: threading.Event):
-        super().__init__()
-
-        self.motor = motor
-        self.signal = Signal()
-        self.started = self.signal.started
-        self.progress = self.signal.progress
-        self.finished = self.signal.finished
-        self._stop_initiated = False
-
-        self.event_to_clear = event_to_clear
-
-    """
-    I ran into an error where I believe the program was writing two
-    messages to the port at the same time (get position, and stop). So,
-    it's important to enforce sequential writing to the port. I'm doing that
-    by putting the stop command in the run loop.
-    """
-
-    def stop(self):
-        self._stop_initiated
-
-    def run(self):
-
-        # TODO may be broken. May need to read location from hardware
-        while self.motor.is_in_motion():
-            pos = self.motor.pos_um()
-            self.progress.emit(pos)
-            # time.sleep(.001)
-
-        # stop flag has been set to True, and the loop has terminated
-        # clear the event
-        self.event_to_clear.clear()
-
-        pos = self.motor.pos_um()
-        self.progress.emit(pos)
-        self.finished.emit(None)
-
-
-class UpdateSpectrumRunnable(qtc.QRunnable):
-    """Runnable class for the ContinuousUpdate class"""
-
-    def __init__(self, spectrometer: Spectrometer, event_to_clear, event_to_set):
-        super().__init__()
-
-        # this class takes as input the spectrometer which it will
-        # continuously pull the the spectrum from
-        self.spectrometer = spectrometer
-        # also initialize a signal so you can transmit the spectrum to the
-        # main Continuous Update class
-        self.signal = Signal()
-        self.started = self.signal.started
-        self.progress = self.signal.progress
-        self.finished = self.signal.finished
-
-        # initialize stop signal to false
-        self._stop = False
-
-        event_to_clear: threading.Event
-        event_to_set: threading.Event
-        self.event_to_clear = event_to_clear
-        self.event_to_set = event_to_set
-
-    # set stop signal to true
-    def stop(self):
-        self._stop = True
-
-    def run(self):
-        # while stop is false, continuously get the spectrum
-        while not self._stop:
-            # get the spectrum
-            # wavelengths, intensities = self.spectrometer.spectrum()
-            spectrum = self.spectrometer.spectrum()
-            # emit the spectrum as a signal
-            # self.progress.emit([wavelengths, intensities])
-            self.progress.emit(spectrum)
-
-        # stop flag has been set to True, and the loop has terminated
-        # clear the event
-        self.event_to_clear.clear()
-        self.event_to_set.set()
 
 
 if __name__ == "__main__":
